Move packages to better location
