+++ src/com/maddyhome/idea/vim/action/motion/updown/MotionShiftDownAction.kt
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.maddyhome.idea.vim.action.motion.leftright
+package com.maddyhome.idea.vim.action.motion.updown
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Editor
